requirements.txt
# For data aquisition
pyserial
aiobmsble
# For dataset creation and management
h5py
dvc
dvc[gdrive]
# For environment variable management
python-dotenv